pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phishing-analyzer"
version = "1.0.0"
description = "Heuristic phishing URL analyzer for SOC/DFIR workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Christian Huhn", email = "duathron@gmail.com" },
]
keywords = ["phishing", "url-analysis", "security", "soc", "dfir", "homoglyph", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.30.0", "openai>=1.30.0"]
dev = ["pytest>=8.0.0", "ruff>=0.4.0"]

[project.scripts]
phishing-analyzer = "phishing_analyzer.main:main"

[project.urls]
Homepage = "https://github.com/duathron/phishing-analyzer"
Repository = "https://github.com/duathron/phishing-analyzer"
"Bug Tracker" = "https://github.com/duathron/phishing-analyzer/issues"

[tool.setuptools.packages.find]
include = ["phishing_analyzer*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
